Feat: 아이템 민팅 기능 추가

+++ src/collections/collections.service.ts
@@ -101,9 +101,9 @@ export class CollectionsService {
     // 컬렉션 삭제
     async deleteCollection(id: number, address: string) {
         const exisCollection = await this.findByOneCollection(id);
-        // if (exisCollection.user.address !== address) {
-        //     throw new NotFoundException(`본인만 수정 가능합니다.`);
-        // }
+        if (exisCollection.user.address !== address) {
+            throw new NotFoundException(`본인만 수정 가능합니다.`);
+        }
         await this.collectionRepository.delete(exisCollection);
     }
 }
